feat: reusable context
